feat(life-event): applied code review recommendation
