+++ src/app/api/clients/[id]/life-events/route.ts
@@ -1,12 +1,6 @@
 import { getDb } from "@/server/db";
-import {
-  client,
-  asset,
-  debt,
-  policy,
-  lifeEventRecalculation,
-} from "@/server/db/schemas";
-import { and, desc, eq, isNull, sql } from "drizzle-orm";
+import { lifeEventRecalculation } from "@/server/db/schemas";
+import { and, desc, eq } from "drizzle-orm";
 import { NextResponse } from "next/server";
 import { z } from "zod";
 import {
@@ -14,13 +8,7 @@ import {
   parseJsonBody,
   handleValidationError,
 } from "@/lib/api/route-helpers";
-import {
-  calculateInsuranceNeedsRounded,
-  DEFAULT_ESTATE_BUFFER,
-  type InsuranceNeedsInput,
-} from "@/lib/financial/insurance-needs";
-import { resolveExistingCoverage } from "@/lib/policy-utils";
-import type { InsuranceNeedsSnapshot } from "@/server/db/schemas/life-events-schema";
+import { computeCurrentSnapshot } from "@/lib/financial/compute-snapshot";
 
 // ============================================================================
 // VALIDATION SCHEMAS
@@ -50,101 +38,6 @@ const triggerLifeEventSchema = z.object({
   /** The current estimate snapshot the advisor sees (before state) */
   beforeSnapshot: insuranceNeedsSnapshotSchema,
 });
-
-// ============================================================================
-// HELPERS
-// ============================================================================
-
-function decimalToNumber(value: string | null | undefined): number {
-  if (value === null || value === undefined) return 0;
-  const num = parseFloat(value);
-  return isNaN(num) ? 0 : num;
-}
-
-/**
- * Compute a fresh insurance needs snapshot from current client DB state.
- * Mirrors the logic in /api/clients/[id]/calculate but returns only snapshot fields.
- */
-async function computeCurrentSnapshot(
-  clientId: string,
-  userId: string,
-): Promise<InsuranceNeedsSnapshot | null> {
-  const db = getDb();
-
-  const clientData = await db.query.client.findFirst({
-    where: and(
-      eq(client.id, clientId),
-      eq(client.userId, userId),
-      isNull(client.deletedAt),
-    ),
-  });
-
-  if (!clientData) return null;
-
-  const [assetTotals] = await db
-    .select({
-      totalAssets: sql<string>`COALESCE(SUM(${asset.currentValue}), 0)`,
-      liquidAssets: sql<string>`COALESCE(SUM(CASE WHEN ${asset.isLiquid} THEN ${asset.currentValue} ELSE 0 END), 0)`,
-    })
-    .from(asset)
-    .where(and(eq(asset.clientId, clientId), isNull(asset.deletedAt)));
-
-  const [debtTotals] = await db
-    .select({
-      totalDebts: sql<string>`COALESCE(SUM(${debt.currentBalance}), 0)`,
-    })
-    .from(debt)
-    .where(and(eq(debt.clientId, clientId), isNull(debt.deletedAt)));
-
-  const [policyTotals] = await db
-    .select({
-      totalActivePolicyCoverage: sql<string>`COALESCE(SUM(CASE WHEN ${policy.status} = 'active' THEN ${policy.faceAmount} ELSE 0 END), 0)`,
-      totalPolicyCount: sql<number>`COUNT(*)`,
-    })
-    .from(policy)
-    .where(and(eq(policy.clientId, clientId), isNull(policy.deletedAt)));
-
-  const totalAssets = decimalToNumber(assetTotals?.totalAssets);
-  const liquidAssets = decimalToNumber(assetTotals?.liquidAssets);
-  const totalDebts = decimalToNumber(debtTotals?.totalDebts);
-  const policyCount = Number(policyTotals?.totalPolicyCount ?? 0);
-  const totalActivePolicyCoverage = decimalToNumber(
-    policyTotals?.totalActivePolicyCoverage,
-  );
-
-  const { existingCoverage } = resolveExistingCoverage({
-    totalPolicyCount: policyCount,
-    activePolicyCoverage: totalActivePolicyCoverage,
-    legacyCoverage: decimalToNumber(clientData.existingLifeInsuranceCoverage),
-  });
-
-  const input: InsuranceNeedsInput = {
-    clientIncome: decimalToNumber(clientData.clientIncome),
-    spouseIncome: decimalToNumber(clientData.spouseIncome),
-    includeSpouseIncome: clientData.hasSpouse ?? false,
-    incomeReplacementPercent: decimalToNumber(
-      clientData.incomeReplacementPercent,
-    ),
-    replacementDurationYears: clientData.replacementDurationYears ?? 10,
-    existingLifeInsuranceCoverage: existingCoverage,
-    totalDebts,
-    liquidAssets,
-    totalAssets,
-    estateBuffer: DEFAULT_ESTATE_BUFFER,
-  };
-
-  const result = calculateInsuranceNeedsRounded(input);
-
-  return {
-    incomeReplacementNeeds: result.incomeReplacementNeeds,
-    debtPayoffNeeds: result.debtPayoffNeeds,
-    estateBufferNeeds: result.estateBufferNeeds,
-    grossNeeds: result.grossNeeds,
-    existingCoverage: result.existingCoverage,
-    liquidAssets: result.liquidAssets,
-    totalInsuranceNeeds: result.totalInsuranceNeeds,
-  };
-}
 
 // ============================================================================
 // GET /api/clients/[id]/life-events
